feat(nutritionist): add NutritionistHome, NutritionistRegister, and SearchClient screens
- Implement NutritionistHomeScreen to display linked clients and allow searching for new clients.
- Create NutritionistRegisterScreen for nutritionist account registration with validation.
- Add SearchClientScreen for searching clients by email and sending link requests.
- Introduce nutritionist service functions for client management and profile creation.
- Update MealDetailScreen to include meal time in meal data.
- Enhance MealResultScreen to calculate average nutritional values per food item.
- Add new API functions for describing meals and updating user profiles.

+++ backend/src/infrastructure/database/typeorm/entities/meal.schema.ts
@@ -35,16 +35,16 @@ export class MealSchema {
   @Column({ type: 'simple-array' })
   foods: string[];
 
-  @Column({ type: 'int' })
+  @Column({ type: 'float' })
   calories: number;
 
-  @Column({ name: 'macros_carbs', type: 'int' })
+  @Column({ name: 'macros_carbs', type: 'float' })
   macrosCarbs: number;
 
-  @Column({ name: 'macros_protein', type: 'int' })
+  @Column({ name: 'macros_protein', type: 'float' })
   macrosProtein: number;
 
-  @Column({ name: 'macros_fat', type: 'int' })
+  @Column({ name: 'macros_fat', type: 'float' })
   macrosFat: number;
 
   @Column({ name: 'meal_time', type: 'timestamptz' })
